# gateway/gateway-controller/api/management-openapi.yaml
openapi: 3.0.3
info:
  title: Gateway Controller Management API
  description: |
    REST API for managing API configurations in the WSO2 API Platform Gateway.
  version: 1.0.0
  contact:
    name: WSO2 API Platform Team
servers:
  - url: http://localhost:9090/api/management/v1
    description: Local development
  - url: http://gateway-controller:9090/api/management/v1
    description: Docker/Kubernetes deployment
security:
  - basicAuth: []
paths:
  /rest-apis:
    post:
      summary: Create a new RestAPI
      description: Add a new RestAPI to the Gateway.
      operationId: createRestAPI
      x-basicauth-roles: [admin, developer]
      tags:
        - Rest API Management
      requestBody:
        required: true
        content:
          application/yaml:
            schema:
              $ref: "#/components/schemas/RestAPIRequest"
          application/json:
            schema:
              $ref: "#/components/schemas/RestAPIRequest"
      responses:
        "201":
          description: RestAPI created successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RestAPI"
        "400":
          description: Invalid configuration (validation failed)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Conflict - API with same name and version already exists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    get:
      summary: List all RestAPIs
      description: List RestAPIs registered in the Gateway, optionally filtered by name, version, context, or status.
      operationId: listRestAPIs
      x-basicauth-roles: [admin, developer]
      tags:
        - Rest API Management
      parameters:
        - name: displayName
          in: query
          required: false
          description: Filter by API display name
          schema:
            type: string
          example: Reading List API
        - name: version
          in: query
          required: false
          description: Filter by API version
          schema:
            type: string
          example: v1.0
        - name: context
          in: query
          required: false
          description: Filter by API context/path
          schema:
            type: string
          example: /reading-list
        - name: status
          in: query
          required: false
          description: Filter by deployment status
          schema:
            type: string
            enum: [ deployed, undeployed ]
          example: undeployed
      responses:
        "200":
          description: List of RestAPIs
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: success
                  count:
                    type: integer
                    example: 5
                  apis:
                    type: array
                    items:
                      $ref: "#/components/schemas/RestAPI"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /rest-apis/{id}:
    get:
      summary: Get RestAPI by id
      description: Get a RestAPI by its ID.
      operationId: getRestAPIById
      x-basicauth-roles: [admin, developer]
      tags:
        - Rest API Management
      parameters:
        - name: id
          in: path
          required: true
          description: |
            Unique public identifier for the API.
          schema:
            type: string
          example: reading-list-api-v1.0
      responses:
        "200":
          description: RestAPI details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RestAPI"
            application/yaml:
              schema:
                $ref: "#/components/schemas/RestAPI"
        "404":
          description: RestAPI not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    put:
      summary: Update an existing RestAPI
      description: Update an existing RestAPI in the Gateway.
      operationId: updateRestAPI
      x-basicauth-roles: [admin, developer]
      tags:
        - Rest API Management
      parameters:
        - name: id
          in: path
          required: true
          description: |
            Unique public identifier of the API to update.
          schema:
            type: string
          example: reading-list-api-v1.0
      requestBody:
        required: true
        content:
          application/yaml:
            schema:
              $ref: "#/components/schemas/RestAPIRequest"
          application/json:
            schema:
              $ref: "#/components/schemas/RestAPIRequest"
      responses:
        "200":
          description: RestAPI updated successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RestAPI"
        "400":
          description: Invalid configuration (validation failed)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: RestAPI not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    delete:
      summary: Delete a RestAPI
      description: Delete a RestAPI from the Gateway.
      operationId: deleteRestAPI
      x-basicauth-roles: [admin, developer]
      tags:
        - Rest API Management
      parameters:
        - name: id
          in: path
          required: true
          description: |
            Unique public identifier of the API to delete.
          schema:
            type: string
          example: reading-list-api-v1.0
      responses:
        "200":
          description: RestAPI deleted successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: success
                  message:
                    type: string
                    example: RestAPI deleted successfully
                  id:
                    type: string
                    example: reading-list-api-v1.0
        "404":
          description: RestAPI not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  securitySchemes:
    basicAuth:
      type: http
      scheme: basic
  schemas:
...
    ResourceStatus:
      type: object
      description: Server-managed lifecycle information for a resource
      properties:
        id:
          type: string
          description: Unique identifier assigned by the server (equal to metadata.name)
          example: reading-list-api-v1.0
        state:
          type: string
          description: Desired deployment state reported by the server
          enum:
            - deployed
            - undeployed
          example: deployed
        createdAt:
          type: string
          format: date-time
          description: Timestamp when the resource was first created (UTC)
          example: 2026-04-24T07:21:13Z
        updatedAt:
          type: string
          format: date-time
          description: Timestamp when the resource was last updated (UTC)
          example: 2026-04-24T07:21:13Z
        deployedAt:
          type: string
          format: date-time
          description: Timestamp when the resource was last deployed (omitted when undeployed)
          example: 2026-04-24T07:21:13Z
        warnings:
          type: array
          readOnly: true
          description: >-
            Non-fatal deploy-time findings (e.g. a policy's parameters resolved
            to something other than what was written, such as an omitted
            accept list inheriting the whole client-CA pool). Present only
            when non-empty.
          items:
            $ref: "#/components/schemas/Warning"
...
    RestAPIRequest:
      type: object
      required:
        - apiVersion
        - metadata
        - kind
        - spec
      properties:
        apiVersion:
          type: string
          description: API specification version
          example: gateway.api-platform.wso2.com/v1
          enum:
            - gateway.api-platform.wso2.com/v1
        kind:
          type: string
          description: API type
          example: RestApi
          enum:
            - RestApi
        metadata:
          $ref: "#/components/schemas/Metadata"
        spec:
          $ref: '#/components/schemas/APIConfigData'
      example:
        apiVersion: gateway.api-platform.wso2.com/v1
        kind: RestApi
        metadata:
          name: reading-list-api-v1.0
        spec:
          displayName: Reading-List-API
          version: v1.0
          context: /reading-list/$version
          upstream:
            main:
              url: https://apis.bijira.dev/samples/reading-list-api-service/v1.0
          policies:
            - name: set-headers
              version: v1
              params:
                request:
                  headers:
                    - name: x-wso2-apip-gateway-version
                      value: v1.0.0
                response:
                  headers:
                    - name: x-environment
                      value: development
          operations:
            - method: GET
              path: /books
            - method: POST
              path: /books
            - method: GET
              path: /books/{id}
            - method: PUT
              path: /books/{id}
            - method: DELETE
              path: /books/{id}

    # Full resource including server-managed status (list/get responses).
    RestAPI:
      allOf:
        - $ref: '#/components/schemas/RestAPIRequest'
        - type: object
          properties:
            status:
              readOnly: true
              description: Server-managed lifecycle fields. Populated on responses.
              allOf:
                - $ref: '#/components/schemas/ResourceStatus'
      example:
        apiVersion: gateway.api-platform.wso2.com/v1
        kind: RestApi
        metadata:
          name: reading-list-api-v1.0
        spec:
          displayName: Reading-List-API
          version: v1.0
          context: /reading-list/$version
          upstream:
            main:
              url: https://apis.bijira.dev/samples/reading-list-api-service/v1.0
          policies:
            - name: set-headers
              version: v1
              params:
                request:
                  headers:
                    - name: x-wso2-apip-gateway-version
                      value: v1.0.0
                response:
                  headers:
                    - name: x-environment
                      value: development
          operations:
            - method: GET
              path: /books
            - method: POST
              path: /books
            - method: GET
              path: /books/{id}
            - method: PUT
              path: /books/{id}
            - method: DELETE
              path: /books/{id}
        status:
          id: reading-list-api-v1.0
          state: deployed
          createdAt: 2026-04-24T07:21:13Z
          updatedAt: 2026-04-24T07:21:13Z
          deployedAt: 2026-04-24T07:21:13Z
    Metadata:
      type: object
      required:
        - name
      properties:
        name:
          type: string
          description: Unique handle for the resource
          example: reading-list-api-v1.0
        labels:
          type: object
          description: Labels are key-value pairs for organizing and selecting APIs. Keys must not contain spaces.
          additionalProperties:
            type: string
          example:
            environment: production
            team: backend
            version: v1
        annotations:
          type: object
          description: Annotations are arbitrary non-identifying metadata. Use domain-prefixed keys.
          additionalProperties:
            type: string
          example:
            gateway.api-platform.wso2.com/project-id: 019d953f-d386-7a64-aa92-1869a28292e0
    APIConfigData:
      type: object
      required:
        - displayName
        - version
        - context
        - upstream
        - operations
        # - kind
      properties:
        displayName:
          type: string
          description: Human-readable API name (must be URL-friendly - only letters, numbers, spaces, hyphens, underscores, and dots allowed)
          minLength: 1
          maxLength: 100
          pattern: '^[a-zA-Z0-9\-_\. ]+$'
          example: Reading List API
        # kind:
        #   type: string
        #   description: API type
        #   example: RestApi
        #   enum:
        #     - RestApi
        version:
          type: string
          description: Semantic version of the API
          pattern: '^v\d+\.\d+$'
          example: v1.0
        context:
          type: string
          description: Base path for all API routes (must start with /, no trailing slash). Use $version to embed the version in the path (e.g., /reading-list/$version resolves to /reading-list/v1.0).
          pattern: '^\/([a-zA-Z0-9_\-\/]*[^\/])?$'
          minLength: 1
          maxLength: 200
          example: /reading-list/$version
        upstreamDefinitions:
          type: array
          description: List of reusable upstream definitions with optional timeout configurations
          items:
            $ref: "#/components/schemas/UpstreamDefinition"
        upstream:
          type: object
          required:
            - main
          description: API-level upstream configuration
          properties:
            main:
              $ref: "#/components/schemas/Upstream"
            sandbox:
              $ref: "#/components/schemas/Upstream"
        vhosts:
          type: object
          required:
            - main
          description: Custom virtual hosts/domains for the API
          properties:
            main:
              type: string
              description: |
                Custom virtual host(s)/domain(s) for production traffic. One or more hostnames
                separated by ';' — each hostname serves the main upstream (e.g. when a Gateway API
                HTTPRoute attaches to multiple listener hostnames). The first entry is the primary
                vhost. Each hostname may be a wildcard such as *.example.com.
              pattern: '^(\*\.)?[a-zA-Z0-9.\-]+(\s*;\s*(\*\.)?[a-zA-Z0-9.\-]+)*$'
              example: api.example.com;docs.example.com;*.example.com
            sandbox:
              type: string
              description: Custom virtual host/domain for sandbox traffic
              pattern: '^[a-zA-Z0-9\.\-]+$'
              example: sandbox-api.example.com
        subscriptionPlans:
          type: array
          description: List of subscription plan names available for this API
          items:
            type: string
          example: ["Gold", "Silver"]
        policies:
          type: array
          description: List of API-level policies applied to all operations unless overridden
          items:
            $ref: "#/components/schemas/Policy"
        resilience:
          $ref: "#/components/schemas/Resilience"
        operations:
          type: array
          description: List of HTTP operations/routes
          minItems: 1
          items:
            $ref: "#/components/schemas/Operation"
        deploymentState:
          type: string
          description: Desired deployment state - 'deployed' (default) or 'undeployed'. When set to 'undeployed', the API is removed from router traffic but configuration, API keys, and policies are preserved for potential redeployment.
          enum: [deployed, undeployed]
          default: deployed
          example: deployed
    UpstreamDefinition:
      type: object
      required:
        - name
        - upstreams
      description: Reusable upstream configuration with optional timeout and load balancing settings
      properties:
        name:
          type: string
          description: Unique identifier for this upstream definition
          minLength: 1
          maxLength: 100
          pattern: '^[a-zA-Z0-9\-_]+$'
          example: my-upstream-1
        basePath:
          type: string
          description: Base path prefix for all endpoints in this upstream (e.g., /api/v2). All requests to this upstream will have this path prepended. Must start with '/' and must not end with '/'; omit for root.
          pattern: '^/[a-zA-Z0-9\-._~!$&''()*+,;=:@%/]*[^/]$'
          example: /api/v2
        timeout:
          $ref: "#/components/schemas/UpstreamTimeout"
        upstreams:
          type: array
          description: List of backend targets with optional weights for load balancing
          minItems: 1
          items:
            type: object
            required:
              - url
            properties:
              url:
                type: string
                format: uri
                description: Backend URL (host and port only, path comes from basePath)
                example: http://prod-backend-1:5000
              weight:
                type: integer
                description: Relative weight for load balancing across multiple upstream targets. Reserved for future multi-target load balancing; not applied yet (only the first target is currently used).
                minimum: 0
                maximum: 100
                example: 80
        tls:
          type: object
          description: >
            Optional mutual-TLS presentation and trust for this backend.
            `identity` (string) names a gateway identity — a certificate
            uploaded to POST /certificates with usage: identity — to
            present on the connection. `trustedCAs` (array of string)
            lists usage: upstream certificates to trust for this backend
            in place of the gateway-wide trust bundle; omit to use the
            gateway bundle. `verifyHostName` (boolean, default true)
            controls SAN/hostname verification against the target. Every
            target in a definition carrying `tls` must use https://.
            Unknown keys are rejected at deploy time.
          additionalProperties: true
          example:
            identity: partner-identity
            trustedCAs: [partner-ca]
            verifyHostName: true
...
    Policy:
      type: object
      required:
        - name
        - version
      properties:
        name:
          type: string
          description: Name of the policy
          example: cors
        version:
          type: string
          description: >
            Version of the policy. Only major-only version is allowed (e.g., v0, v1).
            Full semantic version (e.g., v1.0.0) is not accepted and will be rejected.
            The Gateway Controller resolves the major version to the single matching
            full version installed in the gateway image.
          pattern: '^v\d+$'
          example: v1
        executionCondition:
          type: string
          description: Expression controlling conditional execution of the policy
          example: "request.metadata[authenticated] != true"
        params:
          type: object
          description: Arbitrary parameters for the policy (free-form key/value structure)
          additionalProperties: true

    # -----------------------
    # Webhook (Async) API schema
    # -----------------------
...
    ErrorResponse:
      type: object
      required:
        - status
        - message
      properties:
        status:
          type: string
          example: error
        message:
          type: string
          description: High-level error description
          example: Configuration validation failed
        errors:
          type: array
          description: Detailed validation errors
          items:
            $ref: "#/components/schemas/ValidationError"
    ValidationError:
      type: object
      properties:
        field:
          type: string
          description: Field that failed validation
          example: spec.context
        message:
          type: string
          description: Human-readable error message
          example: Context must start with / and cannot end with /
...