# gateway/gateway-controller/pkg/api/handlers/rest_api_handler.go
/*
 * Copyright (c) 2025, WSO2 LLC. (https://www.wso2.com).
 *
 * WSO2 LLC. licenses this file to you under the Apache License,
 * Version 2.0 (the "License"); you may not use this file except
 * in compliance with the License.
 * You may obtain a copy of the License at
 *
 * http://www.apache.org/licenses/LICENSE-2.0
 *
 * Unless required by applicable law or agreed to in writing,
 * software distributed under the License is distributed on an
 * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
 * KIND, either express or implied.  See the License for the
 * specific language governing permissions and limitations
 * under the License.
 */

package handlers

import (
	"errors"
	"fmt"
	"io"
	"log/slog"
	"net/http"
	"strings"
	"time"

	api "github.com/wso2/api-platform/gateway/gateway-controller/pkg/api/management"
	"github.com/wso2/api-platform/gateway/gateway-controller/pkg/api/middleware"
	"github.com/wso2/api-platform/gateway/gateway-controller/pkg/clientca"
	"github.com/wso2/api-platform/gateway/gateway-controller/pkg/metrics"
	"github.com/wso2/api-platform/gateway/gateway-controller/pkg/models"
	"github.com/wso2/api-platform/gateway/gateway-controller/pkg/service/restapi"
	"github.com/wso2/api-platform/gateway/gateway-controller/pkg/storage"
	"github.com/wso2/api-platform/gateway/gateway-controller/pkg/utils"
	"github.com/wso2/api-platform/httpkit/httputil"
)

// RestAPIHandler handles HTTP requests for REST API CRUD operations.
type RestAPIHandler struct {
	service *restapi.RestAPIService
	logger  *slog.Logger

	// pushArtifactUndeploy is the shared control-plane (DP->CP)
	// push hook, wired from APIServer so REST APIs use the same push path as every other
	// artifact kind. Internally gated (origin/connection/enabled) and run asynchronously;
	// they are nil when no control plane is configured (e.g. in unit tests).
	pushArtifactUndeploy func(cfg *models.StoredConfig, log *slog.Logger)
}

// NewRestAPIHandler creates a new RestAPIHandler.
func NewRestAPIHandler(service *restapi.RestAPIService, logger *slog.Logger) *RestAPIHandler {
	return &RestAPIHandler{
		service: service,
		logger:  logger,
	}
}

// CreateRestAPI implements ServerInterface.CreateRestAPI
// (POST /rest-apis)
func (h *RestAPIHandler) CreateRestAPI(w http.ResponseWriter, r *http.Request) {
	startTime := time.Now()
	operation := "create"

	log := middleware.GetLogger(r, h.logger)

	body, err := io.ReadAll(r.Body)
	if err != nil {
		log.Error("Failed to read request body", slog.Any("error", err))
		metrics.APIOperationsTotal.WithLabelValues(operation, "error", "rest_api").Inc()
		metrics.ValidationErrorsTotal.WithLabelValues(operation, "read_body_failed").Inc()
		httputil.WriteJSON(w, http.StatusBadRequest, api.ErrorResponse{
			Status:  "error",
			Message: "Failed to read request body",
		})
		return
	}

	correlationID := middleware.GetCorrelationID(r)

	result, err := h.service.Create(restapi.CreateParams{
		Body:          body,
		ContentType:   r.Header.Get("Content-Type"),
		CorrelationID: correlationID,
		Logger:        log,
	})
	if err != nil {
		log.Error("Failed to deploy API configuration", slog.Any("error", err))
		metrics.APIOperationsTotal.WithLabelValues(operation, "error", "rest_api").Inc()
		h.mapCreateError(w, err)
		return
	}

	metrics.APIOperationsTotal.WithLabelValues(operation, "success", "rest_api").Inc()
	metrics.APIOperationDurationSeconds.WithLabelValues(operation, "rest_api").Observe(time.Since(startTime).Seconds())
	metrics.APIsTotal.WithLabelValues("rest_api", "active").Inc()

	httputil.WriteJSON(w, http.StatusCreated, h.buildDeployResponse(result.StoredConfig.SourceConfiguration, result.StoredConfig))
}

// ListRestAPIs implements ServerInterface.ListRestAPIs
// (GET /rest-apis)
func (h *RestAPIHandler) ListRestAPIs(w http.ResponseWriter, r *http.Request, params api.ListRestAPIsParams) {
	result, err := h.service.List(params)
	if err != nil {
		httputil.WriteJSON(w, http.StatusInternalServerError, api.ErrorResponse{
			Status:  "error",
			Message: "Failed to retrieve API configurations",
		})
		return
	}

	items := make([]any, 0, len(result.Items))
	for _, cfg := range result.Items {
		conf := cfg.SourceConfiguration
		switch ra := conf.(type) {
		case api.RestAPI:
			if resolved, err := cfg.GetContext(); err == nil {
				ra2 := ra
				ra2.Spec.Context = resolved
				conf = ra2
			}
		case *api.RestAPI:
			if ra != nil {
				if resolved, err := cfg.GetContext(); err == nil {
					ra2 := *ra
					ra2.Spec.Context = resolved
					conf = ra2
				}
			}
		}
		items = append(items, buildResourceResponseFromStored(conf, cfg))
	}

	httputil.WriteJSON(w, http.StatusOK, map[string]any{
		"status": "success",
		"count":  len(items),
		"apis":   items,
	})
}

// GetRestAPIById implements ServerInterface.GetRestAPIById
// (GET /rest-apis/{id})
func (h *RestAPIHandler) GetRestAPIById(w http.ResponseWriter, r *http.Request, id string) {
	log := middleware.GetLogger(r, h.logger)

	result, err := h.service.GetByHandle(id)
	if err != nil {
		h.mapGetError(w, log, id, err)
		return
	}

	cfg := result.Config
	httputil.WriteJSON(w, http.StatusOK, buildResourceResponseFromStored(cfg.SourceConfiguration, cfg))
}

// UpdateRestAPI implements ServerInterface.UpdateRestAPI
// (PUT /rest-apis/{id})
func (h *RestAPIHandler) UpdateRestAPI(w http.ResponseWriter, r *http.Request, id string) {
	startTime := time.Now()
	operation := "update"

	log := middleware.GetLogger(r, h.logger)

	body, err := io.ReadAll(r.Body)
	if err != nil {
		log.Error("Failed to read request body", slog.Any("error", err))
		metrics.APIOperationsTotal.WithLabelValues(operation, "error", "rest_api").Inc()
		metrics.ValidationErrorsTotal.WithLabelValues(operation, "read_body_failed").Inc()
		httputil.WriteJSON(w, http.StatusBadRequest, api.ErrorResponse{
			Status:  "error",
			Message: "Failed to read request body",
		})
		return
	}

	correlationID := middleware.GetCorrelationID(r)

	result, err := h.service.Update(restapi.UpdateParams{
		Handle:        id,
		Body:          body,
		ContentType:   r.Header.Get("Content-Type"),
		CorrelationID: correlationID,
		Logger:        log,
	})
	if err != nil {
		log.Error("Failed to update API configuration", slog.Any("error", err))
		metrics.APIOperationsTotal.WithLabelValues(operation, "error", "rest_api").Inc()
		h.mapUpdateError(w, id, err)
		return
	}

	metrics.APIOperationsTotal.WithLabelValues(operation, "success", "rest_api").Inc()
	metrics.APIOperationDurationSeconds.WithLabelValues(operation, "rest_api").Observe(time.Since(startTime).Seconds())

	httputil.WriteJSON(w, http.StatusOK, h.buildDeployResponse(result.Config.SourceConfiguration, result.Config))
}

// buildDeployResponse builds the create/update response body. For a RestAPI
// SourceConfiguration it resolves mtls-auth's `accept` echo and attaches any
// warnings (see restapi.RestAPIService.ResolveMtlsAuthForResponse); every
// other configuration shape falls back to the plain resource response.
func (h *RestAPIHandler) buildDeployResponse(sourceConfig any, stored *models.StoredConfig) any {
	switch cfg := sourceConfig.(type) {
	case api.RestAPI:
		resolved, warnings := h.service.ResolveMtlsAuthForResponse(cfg)
		warnings = append(warnings, h.service.ResolveUpstreamTLSWarnings(resolved)...)
		h.logDeployWarnings(stored, warnings)
		return buildRestAPIResourceResponseWithWarnings(resolved, stored, warnings)
	case *api.RestAPI:
		if cfg == nil {
			return buildResourceResponseFromStored(sourceConfig, stored)
		}
		resolved, warnings := h.service.ResolveMtlsAuthForResponse(*cfg)
		warnings = append(warnings, h.service.ResolveUpstreamTLSWarnings(resolved)...)
		h.logDeployWarnings(stored, warnings)
		return buildRestAPIResourceResponseWithWarnings(resolved, stored, warnings)
	default:
		return buildResourceResponseFromStored(sourceConfig, stored)
	}
}

// DeleteRestAPI implements ServerInterface.DeleteRestAPI
// (DELETE /rest-apis/{id})
func (h *RestAPIHandler) DeleteRestAPI(w http.ResponseWriter, r *http.Request, id string) {
	startTime := time.Now()
	operation := "delete"

	log := middleware.GetLogger(r, h.logger)

	correlationID := middleware.GetCorrelationID(r)

	result, err := h.service.Delete(restapi.DeleteParams{
		Handle:        id,
		CorrelationID: correlationID,
		Logger:        log,
	})
	if err != nil {
		metrics.APIOperationsTotal.WithLabelValues(operation, "error", "rest_api").Inc()
		h.mapDeleteError(w, log, id, err)
		return
	}

	metrics.APIOperationsTotal.WithLabelValues(operation, "success", "rest_api").Inc()
	metrics.APIOperationDurationSeconds.WithLabelValues(operation, "rest_api").Observe(time.Since(startTime).Seconds())
	metrics.APIsTotal.WithLabelValues("rest_api", "active").Dec()

	// Notify the control plane (DP->CP) that this artifact was deleted via the shared
	// handler path; it keeps the artifact and marks it undeployed.
	if h.pushArtifactUndeploy != nil {
		h.pushArtifactUndeploy(result.Config, log)
	}

	httputil.WriteJSON(w, http.StatusOK, map[string]any{
		"status":  "success",
		"message": "RestAPI deleted successfully",
		"id":      id,
	})
}

// mapCreateError maps service errors to HTTP responses for Create.
func (h *RestAPIHandler) mapCreateError(w http.ResponseWriter, err error) {
	if mapRenderError(w, "create", err) {
		return
	}

	if storage.IsConflictError(err) {
		httputil.WriteJSON(w, http.StatusConflict, api.ErrorResponse{
			Status:  "error",
			Message: err.Error(),
		})
		return
	}

	var validationErr *utils.ValidationErrorListError
	if errors.As(err, &validationErr) {
		apiErrors := make([]api.ValidationError, len(validationErr.Errors))
		for i, e := range validationErr.Errors {
			apiErrors[i] = api.ValidationError{
				Field:   stringPtr(e.Field),
				Message: stringPtr(e.Message),
			}
		}
		httputil.WriteJSON(w, http.StatusBadRequest, api.ErrorResponse{
			Status:  "error",
			Message: "Configuration validation failed",
			Errors:  &apiErrors,
		})
		return
	}

	if isRestAPICreateBadRequest(err) {
		httputil.WriteJSON(w, http.StatusBadRequest, api.ErrorResponse{
			Status:  "error",
			Message: err.Error(),
		})
		return
	}

	httputil.WriteJSON(w, http.StatusInternalServerError, api.ErrorResponse{
		Status:  "error",
		Message: "Failed to create configuration",
	})
}

// mapGetError maps service errors to HTTP responses for GetByHandle.
func (h *RestAPIHandler) mapGetError(w http.ResponseWriter, log *slog.Logger, handle string, err error) {
	log.Warn("API configuration not found", slog.String("handle", handle))
	httputil.WriteJSON(w, http.StatusNotFound, api.ErrorResponse{
		Status:  "error",
		Message: fmt.Sprintf("RestAPI with handle '%s' not found", handle),
	})
}

// mapUpdateError maps service errors to HTTP responses for Update.
func (h *RestAPIHandler) mapUpdateError(w http.ResponseWriter, handle string, err error) {
	if mapRenderError(w, "update", err) {
		return
	}

	var parseErr *restapi.ParseError
	if errors.As(err, &parseErr) {
		metrics.ValidationErrorsTotal.WithLabelValues("update", "parse_failed").Inc()
		httputil.WriteJSON(w, http.StatusBadRequest, api.ErrorResponse{
			Status:  "error",
			Message: fmt.Sprintf("Failed to parse configuration: %v", parseErr.Cause),
		})
		return
	}

	var handleErr *restapi.HandleMismatchError
	if errors.As(err, &handleErr) {
		metrics.ValidationErrorsTotal.WithLabelValues("update", "handle_mismatch").Inc()
		httputil.WriteJSON(w, http.StatusBadRequest, api.ErrorResponse{
			Status:  "error",
			Message: handleErr.Error(),
		})
		return
	}

	var validationErr *restapi.ValidationError
	if errors.As(err, &validationErr) {
		metrics.ValidationErrorsTotal.WithLabelValues("update", "validation_failed").Add(float64(len(validationErr.Errors)))
		apiErrors := make([]api.ValidationError, len(validationErr.Errors))
		for i, e := range validationErr.Errors {
			apiErrors[i] = api.ValidationError{
				Field:   stringPtr(e.Field),
				Message: stringPtr(e.Message),
			}
		}
		httputil.WriteJSON(w, http.StatusBadRequest, api.ErrorResponse{
			Status:  "error",
			Message: "Configuration validation failed",
			Errors:  &apiErrors,
		})
		return
	}

	if errors.Is(err, restapi.ErrNotFound) {
		httputil.WriteJSON(w, http.StatusNotFound, api.ErrorResponse{
			Status:  "error",
			Message: fmt.Sprintf("RestAPI with handle '%s' not found", handle),
		})
		return
	}

	if storage.IsConflictError(err) {
		httputil.WriteJSON(w, http.StatusConflict, api.ErrorResponse{
			Status:  "error",
			Message: err.Error(),
		})
		return
	}

	httputil.WriteJSON(w, http.StatusInternalServerError, api.ErrorResponse{
		Status:  "error",
		Message: "Failed to update configuration",
	})
}

// mapDeleteError maps service errors to HTTP responses for Delete.
func (h *RestAPIHandler) mapDeleteError(w http.ResponseWriter, log *slog.Logger, handle string, err error) {
	if errors.Is(err, restapi.ErrNotFound) {
		log.Warn("API configuration not found", slog.String("handle", handle))
		httputil.WriteJSON(w, http.StatusNotFound, api.ErrorResponse{
			Status:  "error",
			Message: fmt.Sprintf("RestAPI with handle '%s' not found", handle),
		})
		return
	}

	// Topic lifecycle or internal errors
	httputil.WriteJSON(w, http.StatusInternalServerError, api.ErrorResponse{
		Status:  "error",
		Message: "Failed to delete configuration",
	})
}

func isRestAPICreateBadRequest(err error) bool {
	if err == nil {
		return false
	}

	message := strings.ToLower(err.Error())
	return strings.Contains(message, "failed to parse configuration") ||
		strings.Contains(message, "resource kind is required") ||
		strings.Contains(message, "unsupported resource kind") ||
		strings.Contains(message, "invalid or missing origin")
}

// logDeployWarnings records every deploy-time warning at WARN, so a
// condition the response body reports to the caller is also visible to the
// operator reading the controller log.
func (h *RestAPIHandler) logDeployWarnings(stored *models.StoredConfig, warnings []clientca.Warning) {
	if h.logger == nil {
		return
	}
	for _, w := range warnings {
		attrs := []any{slog.String("code", w.Code), slog.String("field", w.Field), slog.String("message", w.Message)}
		if stored != nil {
			attrs = append(attrs, slog.String("api", stored.UUID))
		}
		h.logger.Warn("Deployment warning", attrs...)
	}
}